Find the larger of 2 values by using an if-expressions
There are 2 solutions, because Rust allows the LAST expression in a
function to be the return value of a function.
Expressions do NOT need the semicolon, only statements need that.

However, if the return keyword is used, the expression following it is
returned, and the line is treated as a statement, meaning a semicolon
is needed.