Define a boolean using a let binding. No type annotation required
Rust has a type inference system, like all other Functional
Programming languages.
This means that, depending on the complexity of the expression, Rust's
compiler will be able to determine the type of an expression during
compilation.
This is a huge benefit for writing Rust code, as it means many type
annotations can be left out of certain portions of the code.