Define a color as a struct and as a (struct) tuple
The second version, the (struct) tuple has struct in parentheses,
because the only way to define a type that itself is a tuple is by
defining a struct that is itself a struct.

Inherently, tuples and structs are stored the SAME way.

However, a struct allows you to NAME the fields inside the struct, and
address the various pieces of content by their name.
This may ease the use, development, and reading of the program and its
code.

A tuple uses indexing to reach various elements in a tuple. This is
better for closely related pieces of information that build off each
other to achieve a greater whole, and whose names matter little.
For example, a 3D coordinate point (x, y, z) is likely to be more
easily understood as a tuple than a pure struct.