A string like this "example" is a reference to the underlying heap characters
This means that we actually have a pointer into the **heap** for where
the characters are being stored. We are **not** dealing with the
string struct that is being stored on the stack.
To work with the String struct stored on the stack, we can use the
to_string() function.