Defined the variants of the Message enum
In the case of an enum, something of that enum type can only take ONE
of those values at a time; and because Rust has immutable variables by
default, this means that most of the time, a variable will be of that
variant for its whole life.

A variant is selected by using the Enum as a namespacing scope
delimiter.
This means that to create a variant of an enum, it behaves like a
Java/OOP language static function.