Pattern matching is possible in Rust, usually over enums
This allows us to concisely write a functional if-else if-else type
statement without having a ton of boilerplate code.
Instead, we explicitly state what we are matching over, including the
variant type it is, and we can perform some action based on that.

We can also NOT use the contained values by replacing the name bound
to that value with _. The underscore, _, is the usual "don't care"
value name.