Propagating errors can be done with ?, or a match expression
Because the match construct is actually an expression, we can return
it, just like any other expression.
However, this EXACT same behavior can be achieved with the ? operator.
This operator will attempt to use the value, and if the value is
actually something that is part of Error (this includes Result::Err),
then the function exits early and propagates the error to the calling function.