Functions that use ? MUST have a supported return type
This return type must yield something that can potentially yield
something of type Error.
This means that in this example, the main function can potentially
return an Ok that wraps either a String or an Err that wraps a
ParseIntError.

Because the ? operator requires the function to return something of
those types, and the last expression of a function is ALSO its return
type, the main function must be modified here to format a String
output that is the same thing as the original println!.