Box allows us to have multiple kinds of things
In reality, Box provides us a "box" on the stack that points to
something on the heap. This thing on the heap is allowed to be
relatively unsized compared to the stack, so we can have different
types of Errors pointed to.

This is useful here because the ? operator will return and end
execution early if an Error condition is generated. In that case, then
the Box is utilized and the Error is returned to the calling function.

However, if we proceed without error, the Result's Ok variant contains
the end value.