if-let and while-let bindings of Option
This is almost a syntactic sugar for a match-expression that returns a
value that can be assigned to.

Here, the variant value is matched against one of several options and
is unwrapped accordingly. Again, the actual value can be discarded
with the _ symbol.
Then, based on the option selected, the body of that clause is
executed and runs.

The while-let behaves similarly.
However, vector.pop() wraps the element that is popped with another
layer of Option<T>. Because our vector contains Option<T>, that means
the element we pop off is of type Option<Option<T>>, so we need to
wrap our value in 2 Some()s, like so Some(Some(value)).