Clippy suggests using an epsilon for floating-point comparison.
Because I completed this course on a NixOS host, I needed to add the
shell.nix file and execute nix-shell.

I opted to go another route and use lorri to make management easier.
In this case, I use lorri shell and it runs MY shell, with all its
configurations, and adds the appropriate paths to have everything run.
In addition, the $PATH variable is updated as needed to make my
program run.
To exit this shell C-d works, like most other REPLs.