Use Box to construct recursive structures
This is a similar conundrum to self-recursive structs in C. In C, the
size of the struct MUST be static at compilation time, so a struct
CANNOT refer to itself infinitely. Instead, a pointer to the struct
must be created, which is statically, constantly sized, and the actual
struct value must be stored elsewhere.

The problem is the same here. Because the enum must have fixed size
variants, a List cannot infinitely self-recurse. The use of Box
essentially solves this problem the same way we do in C, by providing
a statically sized pointer that CAN be compiled, and the actual value
is stored elsewhere.

The small downside of using Box is that a Box must be created for even
simply recursive structures.
Borrowing Haskell syntax, the list [1] would be represented as
List::Cons(1, Box::new(List::Nil)).
However, an empty list does NOT require this.
Overall, the cost to pay is relatively low, but still present.