Share numbers and have an Atomic Reference Count to it
The Atomic Reference Counter, Arc, takes in something standard. In
this case, it is a vector of the first 100 numbers, with values from 0
to 99. We need to consume that value by Arc::new to create an atomic
reference counter for that thing.

Then, the use of clone() on that Arc object will ATOMICALLY increment
the reference counter on this object, preventing the thing from being
"garbage collected" by the memory deallocator until it really should
be done.

EACH child will clone() the Arc, and will automatically decrement the
Arc when the thread exits.