The fill_vec fuunction here needs to borrow the vector here
Because the vector is used later, within the main function, the vector
cannot be transferred to the fill_vec function.
Because of this, the function must create a NEW vector that has the
values pushed on and then returned.

When RETURNING something, ownership is TRANSFERRED. Meaning that when
the fill_vec returns, the newvec to main, **main** owns the newvec as
vec1.
If something is defined to be mutable, ANY and ALL other aliases for
that thing must ALSO be labeled as mutable.