Mutability of an argument must be specified
This again comes back to Rust's philosophy of everything being
immutable by default.
In fact, Rust comes from the opposite direction as C does on this
topic. In Rust, you must specify that an argument is mutable with mut.
However, in C, you specify that an argument is NOT MODIFIED with the
const keyword in a formal argument.