Wrapping a Mutex in an Arc makes things slightly complicated, but concurrent
The reason why Arc is automatically dereferenced is that it
implemented the DeRef trait/typeclass.

We first wrap the JobStatus struct into a Mutex, for safe synchronous
access, then wrap the Mutex in an Arc, to ensure that it is alive for
the lifetime of ALL threads that use the Mutex.

When the child thread clone()s the Arc, the reference count is
incremented.
Once incremented, we can acquire the lock with lock(), which returns a
LockResult, which we need to get around with unwrap().
After the unwrap(), we are in the struct, so we can access the
jobs_completed field.

The same thing happens later.