TryFrom<T> gives us the try_from(t: T) function
This functions VERY similarly to the from(t: T) function declared by
the From<T> trait/typeclass.

HOWEVER, the difference with this one is that that we wrap potential
results in the Result enum. This may be "safer" than From<T>'s
version, but may also require more code to handle safely.